Refactor Quick Search to Spotlight Search Component
- Replaced the Quick Search component with a new UsernamePasswordSpotlight component for enhanced functionality.
- Updated the main app component to utilize the new spotlight search feature.
- Removed the old Quick Search component and its associated HTML file.
- Introduced a new spotlight search component with improved layout and search capabilities.
- Implemented IPC communication for searching and managing results in the spotlight search.
