+++ apps/desktop/src/magnify/app/spotlight-search/username-password/username-password-spotlight.component.html
@@ -1,0 +1,141 @@
+<app-spotlight-search
+  placeholder="Search vault…"
+  [results]="results()"
+  (queryChange)="onQuery($event)"
+  (itemAction)="onItemAction($event)"
+>
+  <!-- Result item template -->
+  <ng-template #resultItem let-cipher let-i="index" let-isSelected="isSelected">
+    <!-- Cipher icon -->
+    <div class="tw-flex tw-size-9 tw-shrink-0 tw-items-center tw-justify-center" aria-hidden="true">
+      @if (cipher.faviconUrl) {
+        <img
+          [src]="cipher.faviconUrl"
+          [alt]=""
+          class="tw-size-6 tw-rounded-md tw-object-contain"
+          (error)="
+            $any($event.target).style.display = 'none';
+            $any($event.target).nextElementSibling.style.display = 'inline'
+          "
+        />
+        <i class="bwi bwi-globe bwi-lg tw-text-muted" style="display: none"></i>
+      } @else {
+        <i class="bwi bwi-globe bwi-lg tw-text-muted"></i>
+      }
+    </div>
+
+    <!-- Cipher info -->
+    <div class="tw-min-w-0 tw-flex-1">
+      <div class="tw-truncate tw-text-base tw-font-semibold tw-text-main">
+        {{ cipher.name }}
+      </div>
+      @if (cipher.username) {
+        <div class="tw-truncate tw-text-xs tw-text-muted">
+          {{ cipher.username }}
+        </div>
+      }
+    </div>
+
+    <!-- Action chips (selected row only) + shortcut badge -->
+    <div class="tw-flex tw-shrink-0 tw-items-center tw-gap-2">
+      @if (isSelected) {
+        @if (isMac) {
+          <!-- macOS: labeled chips -->
+          <div class="tw-flex tw-items-center tw-gap-1.5">
+            <button
+              type="button"
+              aria-label="Copy password"
+              class="tw-flex tw-items-center tw-gap-1 tw-rounded-md tw-bg-primary-100 tw-px-2.5 tw-py-1 tw-text-xs tw-font-medium tw-text-primary-700 tw-transition-colors"
+              (click)="$event.stopPropagation(); copyPassword(cipher)"
+            >
+              Copy password
+              <kbd>↵</kbd>
+            </button>
+            @if (cipher.username) {
+              <button
+                type="button"
+                aria-label="Copy username"
+                class="tw-flex tw-items-center tw-gap-1 tw-rounded-md tw-bg-primary-100 tw-px-2.5 tw-py-1 tw-text-xs tw-font-medium tw-text-primary-700 tw-transition-colors"
+                (click)="$event.stopPropagation(); copyUsername(cipher)"
+              >
+                Copy username
+                <kbd>⌘C</kbd>
+              </button>
+            }
+            @if (cipher.hasUri) {
+              <button
+                bitIconButton="bwi-external-link"
+                size="small"
+                type="button"
+                label="Launch"
+                (click)="$event.stopPropagation(); launchCipher(cipher)"
+              ></button>
+            }
+          </div>
+        } @else {
+          <!-- Non-macOS: icon buttons -->
+          <div class="tw-flex tw-items-center tw-gap-0.5">
+            <button
+              bitIconButton="bwi-key"
+              size="small"
+              type="button"
+              label="Copy password"
+              (click)="$event.stopPropagation(); copyPassword(cipher)"
+            ></button>
+            @if (cipher.username) {
+              <button
+                bitIconButton="bwi-person"
+                size="small"
+                type="button"
+                label="Copy username"
+                (click)="$event.stopPropagation(); copyUsername(cipher)"
+              ></button>
+            }
+            @if (cipher.hasUri) {
+              <button
+                bitIconButton="bwi-external-link"
+                size="small"
+                type="button"
+                label="Launch"
+                (click)="$event.stopPropagation(); launchCipher(cipher)"
+              ></button>
+            }
+          </div>
+        }
+      }
+
+      <!-- ⌘N / ^N shortcut badge -->
+      @if (i < maxResults) {
+        <kbd
+          class="tw-min-w-[2rem] tw-rounded tw-px-1.5 tw-py-0.5 tw-text-center tw-text-xs tw-font-mono"
+          [class.tw-border]="true"
+          [class.tw-border-secondary-300]="!isSelected"
+          [class.tw-border-primary-300]="isSelected"
+          [class.tw-bg-secondary-100]="!isSelected"
+          [class.tw-text-muted]="!isSelected"
+          [class.tw-bg-primary-50]="isSelected"
+          [class.tw-text-primary-600]="isSelected"
+          >{{ isMac ? "⌘" : "^" }}{{ i + 1 }}</kbd
+        >
+      }
+    </div>
+  </ng-template>
+
+  <!-- Footer hints template -->
+  <ng-template #footerHints>
+    <span class="tw-flex tw-items-center tw-gap-1.5">
+      <kbd
+        class="tw-rounded tw-border tw-border-secondary-300 tw-bg-secondary-100 tw-px-1.5 tw-py-0.5 tw-font-mono tw-text-xs"
+        >↵</kbd
+      >
+      Copy password
+    </span>
+    <span class="tw-flex tw-items-center tw-gap-1.5">
+      <kbd
+        class="tw-rounded tw-border tw-border-secondary-300 tw-bg-secondary-100 tw-px-1.5 tw-py-0.5 tw-font-mono tw-text-xs"
+        >{{ isMac ? "⌘C" : "Ctrl+C" }}</kbd
+      >
+      Copy username
+    </span>
+  </ng-template>
+</app-spotlight-search>
